% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gcSummary.R
\name{gcSummary}
\alias{gcSummary}
\title{Google Citation summary}
\usage{
gcSummary(alldata)
}
\arguments{
\item{alldata}{googleCite output}
}
\value{
Printout of outcomes
}
\description{
Google Citation summary
}

