% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getAuthor}
\alias{getAuthor}
\title{Get the author that the googleCite is based on}
\usage{
getAuthor(webpage)
}
\arguments{
\item{webpage}{Google Citation Webpage}
}
\value{
name of authors
}
\description{
Get the author that the googleCite is based on
}

