% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPapers.R
\name{getPapers}
\alias{getPapers}
\title{Grab the papers from the Google Citations page}
\usage{
getPapers(theurl)
}
\arguments{
\item{theurl}{URL to be parsed}
}
\description{
Grab the papers from the Google Citations page
}

