\name{makeAuthorCloud}
\alias{makeAuthorCloud}
\title{Author cloud maker}
\usage{
  makeAuthorCloud(tab, splitter = " ", auth_pat,
    addastopwords = NULL, ...)
}
\arguments{
  \item{tab}{Table generated by googleCite}

  \item{splitter}{split author names (default " ")}

  \item{auth_pat}{Authors to drop}
}
\value{
  NULL (Plot rendered)
}
\description{
  Make word cloud of papers
}

