% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCitePerYear}
\alias{getCitePerYear}
\title{Get citation and h indices per year}
\usage{
getCitePerYear(theurl, year = NULL, url = TRUE,
  this.year = as.numeric(format(Sys.time(), "\%Y")), pagegraph = FALSE)
}
\arguments{
\item{theurl}{url from Google Citation or htmlTreeParse if (url=FALSE)}

\item{year}{year to start from (default NULL)}

\item{url}{(logical) is theurl a url or htmlTree}

\item{this.year}{The end year (default this year)}

\item{pagegraph}{Use this for graphs of an article}
}
\description{
Get citation and h indices per year
}

