% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleCite.R
\name{googleCite}
\alias{googleCite}
\title{Workhorse function to extract information from google
citations page}
\usage{
googleCite(theurl, citations = FALSE, plotIt = FALSE, pdfname = NULL,
  dlpdfs = FALSE, dlfolder = NULL, ...)
}
\arguments{
\item{theurl}{URL of the google Citation page to be analyzed}

\item{citations}{(logical) Obsolete}

\item{plotIt}{Plot the Author/Paper wordcloud}

\item{pdfname}{(NULL) filename of plot (if plotIt = TRUE)}

\item{dlpdfs}{(logical) Beta - download pdfs}

\item{dlfolder}{(file path) folder to download pdf (default getwd())}

\item{...}{parameters to be passed to makeAuthorCloud, makePaperCloud}
}

