% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAuthorCloud.R
\name{makeAuthorCloud}
\alias{makeAuthorCloud}
\title{Author cloud maker}
\usage{
makeAuthorCloud(tab, splitter = ", ", auth_pat = NULL,
  addastopwords = NULL, verbose = TRUE, ...)
}
\arguments{
\item{tab}{Table generated by googleCite}

\item{splitter}{split author names (default " ")}

\item{auth_pat}{Authors to drop}

\item{addastopwords}{words to delete from wordcloud}

\item{verbose}{Diagnostic message}

\item{...}{Not currently used}
}
\value{
NULL (Plot rendered)
}
\description{
Make word cloud of papers
}

