% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makePaperCloud}
\alias{makePaperCloud}
\title{Paper Cloud Maker}
\usage{
makePaperCloud(tab, addstopwords = NULL, ...)
}
\arguments{
\item{tab}{Table generated by googleCite}

\item{addstopwords}{words to remove}

\item{...}{Not currently used.}
}
\value{
NULL (Plot rendered)
}
\description{
Make word cloud of papers
}

